add failing test
